--- a/spec/lib/facilities/nca_facility_spec.rb
+++ b/spec/lib/facilities/nca_facility_spec.rb
@@ -14,7 +14,7 @@ module Facilities
       it 'should pull data from ArcGIS endpoint' do
         VCR.use_cassette('facilities/va/nca_facilities') do
           list = NCAFacility.pull_source_data
-          expect(list.size).to eq(173)
+          expect(list.size).to eq(10)
         end
       end