--- a/spec/lib/facilities/vha_facility_spec.rb
+++ b/spec/lib/facilities/vha_facility_spec.rb
@@ -11,9 +11,9 @@ module Facilities
 
     describe 'pull_source_data' do
       it 'should pull data from ArcGIS endpoint' do
-        VCR.use_cassette('facilities/va/vha_facilities') do
+        VCR.use_cassette('facilities/va/vha_facilities_limit_results') do
           list = VHAFacility.pull_source_data
-          expect(list.size).to eq(1186)
+          expect(list.size).to eq(4)
           expect(list.all? { |item| item.is_a?(VHAFacility) })
         end
       end