# Facility Locator Download Job
## Before
```
rspec spec/jobs/facilities/facility_location_download_job_spec.rb
...........
Finished in 27.57 seconds (files took 7.76 seconds to load)
11 examples, 0 failures
```
## After
```spec spec/jobs/facilities/facility_location_download_job_spec.rb
...........
Finished in 1.73 seconds (files took 7.16 seconds to load)
11 examples, 0 failures
```
# All specs
## Before
```
Finished in 3 minutes 17.3 seconds (files took 10.87 seconds to load)
2541 examples, 0 failures, 2 pending
```
## After
```
Finished in 2 minutes 47.7 seconds (files took 9.96 seconds to load)
2541 examples, 0 failures, 2 pending
```